background_service/requirements.txt
httpx==0.27.2
Pillow>=11.0.0
pystray==0.19.5
